Move app folder to root level
This allows easy access to the crust of the application. Additionally, it allows the cmd packages to import from "/app" instead of from internal packages.
