+++ app/inmem/app.go
@@ -1,8 +1,8 @@
 package inmem
 
 import (
+	"github.com/albertowusuasare/customer-app/app"
 	"github.com/albertowusuasare/customer-app/internal/adding"
-	"github.com/albertowusuasare/customer-app/internal/app"
 	queue "github.com/albertowusuasare/customer-app/internal/msg/inmem"
 	"github.com/albertowusuasare/customer-app/internal/storage/inmem"
 	"github.com/albertowusuasare/customer-app/internal/uuid"
