+++ app/remove.go
