+++ architecture/README.md
@@ -1,7 +1,7 @@
 ## High level
 This application exposes functionality to create, read, update and delete a customer. 
 At a high level, the application is composed of workflows (see /internal/workflow). Each 
-workflow is a vertically complete customer feature. Workflows are pipelines that compose even smaller bits of functionality for example storage, event publishing etc. The entire set of workflows in the application are encapsulated in an 'app.Customer' (see internal/app). 'This encapsulation allows us to create different versions of the app. For instance, we are able to create an 'inmemapp' (see 'internal/app/inmem.go') which uses an inmemory storage mechanism and hence usueful during testing. The application is then wired up in a main function such as in '/cmd/customer-svc/main.go'.
+workflow is a vertically complete customer feature. Workflows are pipelines that compose even smaller bits of functionality for example storage, event publishing etc. The entire set of workflows in the application are encapsulated in an 'app.Customer' (see /app). 'This encapsulation allows us to create different versions of the app. For instance, we are able to create an 'inmemapp' (see '/app/inmem.go') which uses an inmemory storage mechanism and hence usueful during testing. The application is then wired up in a main function such as in '/cmd/customer-svc/main.go'.
 
 <img src="./architecture.svg?sanitize=true">
