+++ cmd/customer-svc-standalone-gcp/main.go
@@ -7,8 +7,8 @@ import (
 	"net/http"
 
 	"cloud.google.com/go/firestore"
-	"github.com/albertowusuasare/customer-app/internal/app"
-	"github.com/albertowusuasare/customer-app/internal/app/google"
+	"github.com/albertowusuasare/customer-app/app"
+	"github.com/albertowusuasare/customer-app/app/google"
 )
 
 func main() {
