+++ test/integration/create_fixture.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 
-	"github.com/albertowusuasare/customer-app/internal/app"
+	"github.com/albertowusuasare/customer-app/app"
 )
 
 // CreateTestDataCustomer returns a testdata customer
